Remove action onClick from BarChart. Add useEffect liesten to route.
