+++ src/components/BarChart/BarChart-redux.tsx
@@ -3,10 +3,8 @@ import { useMappedState, useDispatch } from 'redux-react-hook';
 
 import BarChartContainer from './BarChart-container';
 import { State } from './types';
-import { Wallets, TransactionsSummedData, TransactionsData } from '../../types';
+import { Wallets } from '../../types';
 
-import * as transactionsActionsSummed from '../../store/actions/ethereum/transactions-summed';
-import * as transactionsActions from '../../store/actions/ethereum/transactions';
 import * as walletActions from '../../store/actions/ethereum/wallets';
 
 
@@ -20,37 +18,14 @@ const BarChartRedux = () => {
 
     const { wallets } = useMappedState(mapState);
 
-    const fetchEthereumTransactionsSummed = (transactionsSummedData: TransactionsSummedData): void => {
-        dispatch({
-            type: transactionsActionsSummed.ETHEREUM_FETCH_TRANSACTIONS_SUMMED,
-            transactionsSummedData: transactionsSummedData
-        });
-    };
-    
-    const fetchEthereumTransactions = (transactionsData: TransactionsData): void => {
-        dispatch({
-            type: transactionsActions.ETHEREUM_FETCH_TRANSACTIONS,
-            transactionsData: transactionsData
-        });
-    };
-
     const fetchEthereumWallets = (): void => {
         dispatch({
           type: walletActions.ETHEREUM_FETCH_WALLETS
         });
       };
 
-    const flushEthereumTransactions = (): void => {
-        dispatch({
-          type: transactionsActions.ETHEREUM_FLUSH_TRANSACTIONS
-        });
-      };
-
     const actions = {
-        fetchEthereumTransactionsSummed,
-        fetchEthereumWallets,
-        fetchEthereumTransactions,
-        flushEthereumTransactions
+        fetchEthereumWallets
     }
 
     return (
