+++ src/components/LineChart/LineChart-container.tsx
@@ -17,6 +17,9 @@ const LineCharts = (props: LineChartProps): React.ReactElement => {
   ]);
   const [data, setData] = useState([10, 20, 30, 40]);
 
+  let route = props.match.url;
+
+
   const checkWalletHashAndFetchTransactions = () => {
     if (walletHash) {
       actions.fetchEthereumTransactionsSummed({walletHash:walletHash});
@@ -42,7 +45,7 @@ const LineCharts = (props: LineChartProps): React.ReactElement => {
 
   useEffect((): void => {
     checkWalletHashAndFetchTransactions();
-  }, []);
+  }, [route]);
 
 
   return (
