+++ src/components/TransactionsList/TransactionsList-styles.tsx
@@ -1,26 +1,6 @@
 import { makeStyles } from '@material-ui/core/styles';
 
-export const transactionsDetailsModalStyle = makeStyles(theme => ({
-  dialogTitle: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    '& h2': {
-      width: '100%',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'space-between'
-    }
-  },
-  dialogAction: {
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(2),
-    paddingRight: theme.spacing(3),
-    paddingLeft: theme.spacing(3)
-  }
-}));
-
-export const transactionsListTableStyle = makeStyles( () => ({
+export const useTransactionsListTableStyles = makeStyles( () => ({
   thead: {
     display: 'block'
   }, 
@@ -36,8 +16,4 @@ export const transactionsListTableStyle = makeStyles( () => ({
   td: {
     width: '24vw'
   }
-
 }));
-
-
-
